refactor(#615): one truncation and one relative-time implementation crate-wide
- card_parse::truncate_chars_with(s, max, suffix) generalizes the existing
  truncate_chars (which becomes the "..." form) and carries the #346
  underflow guard for any suffix.
- status::truncate keeps its first-line semantics as a thin wrapper over
  truncate_chars; its cap now includes the ellipsis (was max+3 bytes).
- usage::truncate_str keeps its "~" column marker as a thin wrapper over
  truncate_chars_with; same cap, plus the underflow guard its copy lacked.
- New timefmt module owns relative_time (promoted verbatim from status.rs,
  tests moved with it); usage::format_start_time delegates to it, deleting
  the hand-rolled ts[11..13]/ts[14..16] clock math that only handled
  same-day timestamps and silently clamped negatives. Non-today sessions
  now render as "Nd ago" instead of a truncated date prefix.
